refactor(pi-fence): move tree-sitter logic to tree-sitter.ts, rename loadParser
- New src/tree-sitter.ts: resolveTreeSitterWasm, loadTreeSitterParser
  (renamed from loadParser), extractTreeSitterNodes, collectNodes,
  ensureInitialized, parserCache — everything that touches web-tree-sitter
- src/parse.ts: now only CommentNode interface and getCommentHash
- All language definitions updated to import from ../tree-sitter.js
