+++ packages/pi-fence/src/languages/css.ts
@@ -1,4 +1,8 @@
-import { extractTreeSitterNodes, loadParser, resolveTreeSitterWasm } from "../parse.js";
+import {
+  extractTreeSitterNodes,
+  loadTreeSitterParser,
+  resolveTreeSitterWasm,
+} from "../tree-sitter.js";
 import type { LanguageDefinition } from "./index.js";
 
 const wasmPath = resolveTreeSitterWasm("tree-sitter-css", "tree-sitter-css.wasm");
@@ -6,7 +10,7 @@ const wasmPath = resolveTreeSitterWasm("tree-sitter-css", "tree-sitter-css.wasm"
 export const css: LanguageDefinition = {
   supportedExtensions: ["css"],
   async extractCommentNodes(content, signal) {
-    const parser = await loadParser(wasmPath, signal);
+    const parser = await loadTreeSitterParser(wasmPath, signal);
     if (!parser) {
       return [];
     }
